The population of a district is added (1 billion is missing.[40478654]

+++ src/main/java/com/napier/sem/City.java
@@ -26,6 +26,6 @@ public class City {
     /**
      * City Population
      */
-    public int Population;
+    public long Population;
 
 }
